Extract page metadata into its own module
