+++ frontend/src/components/layout/Layout.tsx
@@ -5,9 +5,7 @@ import {
   useEffect,
   useState,
 } from "react";
-import Head from "next/head";
 import Link from "next/link";
-import { useRouter } from "next/router";
 import { useLocalization } from "@fluent/react";
 import { ToastContainer, toast, Slide } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
@@ -18,17 +16,15 @@ import logoTypePremiumLight from "../../../../static/images/fx-private-relay-pre
 import logoTypePremiumDark from "../../../../static/images/fx-private-relay-premium-logotype-dark.svg";
 import logo from "../../../../static/images/placeholder-logo.svg";
 import mozillaLogo from "../../../../static/images/logos/moz-logo-bw-rgb.svg";
-import favicon from "../../../public/favicon.svg";
-import socialMediaImage from "../../../../static/images/share-relay.jpg";
 import { useProfiles } from "../../hooks/api/profile";
 import { Navigation } from "./navigation/Navigation";
 import { useIsLoggedIn } from "../../hooks/session";
-import { getRuntimeConfig } from "../../config";
 import { TopMessage } from "./topmessage/TopMessage";
 import { makeToast } from "../../functions/makeToast";
 import { useUsers } from "../../hooks/api/user";
 import { MobileNavigation } from "./navigation/MobileNavigation";
 import { CloseIcon } from "../Icons";
+import { PageMetadata } from "./PageMetadata";
 
 export type Props = {
   children: ReactNode;
@@ -42,7 +38,6 @@ export const Layout = (props: Props) => {
   const { l10n } = useLocalization();
   const profiles = useProfiles();
   const isLoggedIn = useIsLoggedIn();
-  const router = useRouter();
   const hasPremium: boolean = profiles.data?.[0].has_premium ?? false;
   const usersData = useUsers().data?.[0];
   const [mobileMenuExpanded, setMobileMenuExpanded] = useState<boolean>();
@@ -103,33 +98,7 @@ export const Layout = (props: Props) => {
 
   return (
     <>
-      <Head>
-        <link rel="icon" type="image/svg+xml" href={favicon.src}></link>
-        <title>{l10n.getString("meta-title")}</title>
-        <meta
-          name="description"
-          content={l10n.getString("meta-description-2")}
-        />
-        <meta
-          property="og:url"
-          content={getRuntimeConfig().frontendOrigin + router.asPath}
-        />
-        <meta property="og:type" content="website" />
-        <meta property="og:title" content={l10n.getString("meta-title")} />
-        <meta
-          property="og:description"
-          content={l10n.getString("meta-description-2")}
-        />
-        <meta property="og:image" content={socialMediaImage.src} />
-        <meta name="twitter:card" content="summary_large_image" />
-        <meta name="twitter:site" content="@firefox" />
-        <meta name="twitter:title" content={l10n.getString("meta-title")} />
-        <meta
-          name="twitter:description"
-          content={l10n.getString("meta-description-2")}
-        />
-        <meta name="twitter:image" content={socialMediaImage.src} />
-      </Head>
+      <PageMetadata />
       <div className={styles.wrapper}>
         {apiMockWarning}
         <TopMessage profile={profiles.data?.[0]} />
